Try and validate the race condition in run.py
Lets try to wait a few seconds after we drained the queue, and re-check it again.
If it has any items there, it means that we indeed have a real, proved, race condition.

NOTE - this is just a test patch, to try and prove that there is a race condition, not a fix for the issue.

Signed-off-by: Lev Veyde <lveyde@redhat.com>